Fixes #1162 - replaces about page with an Android Fragment
