+++ app/src/main/java/org/mozilla/fenix/AppRequestInterceptor.kt
@@ -9,34 +9,13 @@ import mozilla.components.browser.errorpages.ErrorPages
 import mozilla.components.browser.errorpages.ErrorType
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.request.RequestInterceptor
-import org.mozilla.fenix.ext.components
-import org.mozilla.fenix.settings.AboutPage
-import org.mozilla.fenix.settings.SettingsFragment
 
 class AppRequestInterceptor(private val context: Context) : RequestInterceptor {
-    override fun onLoadRequest(
-        session: EngineSession,
-        uri: String
-    ): RequestInterceptor.InterceptionResponse? {
-        return when (uri) {
-            SettingsFragment.aboutURL -> {
-                val page = AboutPage.createAboutPage(context)
-                return RequestInterceptor.InterceptionResponse.Content(page, encoding = base64)
-            }
-
-            else -> context.components.services.accountsAuthFeature.interceptor.onLoadRequest(session, uri)
-        }
-    }
-
     override fun onErrorRequest(
         session: EngineSession,
         errorType: ErrorType,
         uri: String?
     ): RequestInterceptor.ErrorResponse? {
         return RequestInterceptor.ErrorResponse(ErrorPages.createErrorPage(context, errorType))
-    }
-
-    companion object {
-        const val base64 = "base64"
     }
 }
