--- app/src/main/res/raw/about.html
@@ -1,52 +1,0 @@
-<!doctype html>
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this file,
-   - You can obtain one at http://mozilla.org/MPL/2.0/.  -->
-<html>
-<head>
-    <meta
-            name="viewport"
-            charset="utf-8"
-            content="width=device-width, initial-scale=1">
-    <style>
-body, html {
-    background: #f9f9fa;
-    color: #0C0C0D;
-    font-family: sans-serif;
-    line-height: 24px;
-    font-size: 14px;
-}
-body{
-    padding-left: 24px;
-    padding-right: 24px;
-    margin-left: 0px;
-    margin-right: 0px;
-}
-a {
-    color: #0A9AF4;
-}
-p.subtitle {
-    text-align: center;
-    opacity: .7;
-    margin: 0;
-}
-img#wordmark {
-    /* We need to set the dp size here, because by default webview assumes the image is not
-    density specific (but since it's an android resource, we get a density specific version). */
-    width: 180px;
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    padding-top: 24px;
-    }
-    </style>
-</head>
-<body class="about">
-    <br>
-    <img src="%wordmark%" id="wordmark"/>
-    <br>
-    <p class="subtitle">%about-version%</p>
-    %about-content%
-    <p class="subtitle">Built on: %build-date%</p>
-</body>
-</html>
